requirements.txt
nipype
scipy
numpy
seaborn
matplotlib
nibabel
fmriprep
niworkflows
pybids
pydicom
